agent/pyproject.toml
[tool.poetry]
name = "agent"
version = "0.1.0"
description = ""
authors = ["CopilotKit"]
readme = "README.md"
packages = [{include = "sample_agent"}]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
copilotkit = "0.1.69"
langchain = "0.3.26"
langchain-openai = "0.3.28"
langgraph = "0.6.11"
langgraph-checkpoint-sqlite = "^2.0.10"
langsmith = "0.4.37"
openai = "1.109.1"
fastapi = "^0.115.14"
uvicorn = "^0.38.0"
python-dotenv = "^1.1.1"
aiosqlite = "^0.21.0"

[tool.poetry.scripts]
demo = "sample_agent.demo:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
